+++ StereoPlus2/Renderer.hpp
@@ -18,11 +18,6 @@ class Renderer {
 	const int stencilBufferMaskDim1 = 0x4;
 	const int stencilBufferMaskDim2 = 0x8;
 
-	glm::vec4 defaultColorLeft = glm::vec4(1, 0, 0, 1);
-	glm::vec4 defaultColorRight = glm::vec4(0, 1, 1, 1);
-	glm::vec4 dimmedColorLeft = glm::vec4(1, 0, 0, 0.5);
-	glm::vec4 dimmedColorRight = glm::vec4(0, 1, 1, 0.5);
-
 	glm::vec4 whiteColorBright = glm::vec4(1, 1, 1, 1);
 	glm::vec4 whiteColorDim = glm::vec4(1, 1, 1, 0.5);
 
@@ -97,7 +92,7 @@ class Renderer {
 		ShaderLeft = GLLoader::CreateShaderProgram(vertexShaderSource, fragmentShaderSourceLeft);
 		ShaderRight = GLLoader::CreateShaderProgram(vertexShaderSource, fragmentShaderSourceRight);
 
-		UpdateShaderColor(defaultColorLeft, defaultColorRight);
+		UpdateShaderColor(Settings::ColorLeft().Get(), Settings::ColorRight().Get());
 
 		UpdateShaderColor(whiteSquare.ShaderProgram, whiteColorBright, "myColor");
 		UpdateShaderColor(whiteSquareDim.ShaderProgram, whiteColorDim, "myColor");
@@ -127,7 +122,7 @@ class Renderer {
 	}
 
 	void DrawBright(StereoCamera* camera, SceneObject* o) {
-		UpdateShaderColor(defaultColorLeft, defaultColorRight);
+		UpdateShaderColor(Settings::ColorLeft().Get(), Settings::ColorRight().Get());
 		o->Draw(
 			[&camera](const glm::vec3& p) { return camera->GetLeft(p); },
 			[&camera](const glm::vec3& p) { return camera->GetRight(p); },
@@ -137,7 +132,7 @@ class Renderer {
 			stencilBufferMaskBright2);
 	}
 	void DrawDim(StereoCamera* camera, SceneObject* o) {
-		UpdateShaderColor(dimmedColorLeft, dimmedColorRight);
+		UpdateShaderColor(Settings::DimmedColorLeft().Get(), Settings::DimmedColorRight().Get());
 		o->Draw(
 			[&camera](const glm::vec3& p) { return camera->GetLeft(p); },
 			[&camera](const glm::vec3& p) { return camera->GetRight(p); },